# Simplify and speed up common prefix searcher