--- a/src/trie.rs
+++ b/src/trie.rs
@@ -205,11 +205,11 @@ impl Trie {
     /// let keys = vec!["世界", "世界中", "国民"];
     /// let trie = Trie::from_keys(&keys).unwrap();
     ///
-    /// let haystack: Vec<_> = "国民が世界中にて".chars().collect();
+    /// let haystack: Vec<char> = "国民が世界中にて".chars().collect();
     /// let mut matches = vec![];
     ///
     /// for i in 0..haystack.len() {
-    ///     for (v, j) in trie.common_prefix_search(haystack[i..].iter().cloned()) {
+    ///     for (v, j) in trie.common_prefix_search(haystack[i..].iter().copied()) {
     ///         matches.push((v, i..i + j));
     ///     }
     /// }
@@ -358,7 +358,7 @@ mod tests {
         let mut matches = vec![];
 
         for i in 0..haystack.len() {
-            for (v, j) in trie.common_prefix_search(haystack[i..].iter().cloned()) {
+            for (v, j) in trie.common_prefix_search(haystack[i..].iter().copied()) {
                 matches.push((v, i..i + j));
             }
         }