This PR simplifies the implementation of common prefix search and avoids storing mapped characters in an auxiliary data structure. With this modification, the search speed was improved as follows.

```
unidic/cps/crawdad/trie time:   [6.9536 ms 6.9676 ms 6.9839 ms]
                        change: [-14.553% -13.982% -13.439%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) high mild
unidic/cps/crawdad/mptrie
                        time:   [8.1081 ms 8.1424 ms 8.1795 ms]
                        change: [-18.109% -17.562% -16.983%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
```

## Discussion

This modification contains the following trade off.

 - Pros.
   - Make the API simpler.
   - Remove the auxiliary data structure to store mapped characters.
 - Cons.
   - Map the same characters in a haystack repeatedly.
   - Omit byte positions from matched results.

The experimental results indicate that storing and accessing mapped characters in an array is more costly than mapping characters repeatedly. The new simpler API would be easier to use, since it is not necessary to create an instance to store characters. Byte positions are not a necessary result for all users, so there would be no need to provide them in this library.