@vbkaisetsu 2022-07-24
@kampersanda 
Thank you for the improvements.

I think for most users, byte positions are more valuable than code point positions because they store text as a byte array in most cases.
However, keeping in mind that the argument takes an iterator of chars, I do not think it is a bad idea to keep only code point positions.
I want to give some thought to which would be better.